requirements.txt
python-telegram-bot==20.6
pymongo==4.5.0
openpyxl==3.1.2
python-dotenv
fastapi
uvicorn